Move AgendaItem.php
