+++ database/seeders/AgendaItemSeeder.php
@@ -14,7 +14,7 @@ class AgendaItemSeeder extends Seeder
      */
     public function run()
     {
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(5)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(5)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -27,7 +27,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 1;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(12)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(12)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -40,7 +40,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(7)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(7)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -53,7 +53,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(14)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(14)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -66,7 +66,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(21)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(21)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -79,7 +79,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(21)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(21)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -92,7 +92,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(28)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(28)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -105,7 +105,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(35)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(35)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -118,7 +118,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(42)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(42)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -131,7 +131,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(49)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(49)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
@@ -144,7 +144,7 @@ class AgendaItemSeeder extends Seeder
         $agendaItem->application_form_id = 2;
         $agendaItem->save();
 
-        $agendaItem = new \App\AgendaItem();
+        $agendaItem = new \App\Models\AgendaItem();
         $agendaItem->startDate = Carbon::now()->addDays(56)->addHours(12);
         $agendaItem->endDate = Carbon::now()->addDays(56)->addHours(15);
         $agendaItem->subscription_endDate = Carbon::now()->addDays(2)->addHours(15);
