+++ tests/Feature/Agenda/DeleteAgendaItemTest.php
@@ -2,7 +2,7 @@
 
 namespace Tests\Feature\Agenda;
 
-use App\AgendaItem;
+use App\Models\AgendaItem;
 use App\User;
 use Illuminate\Foundation\Testing\DatabaseMigrations;
 use Illuminate\Support\Facades\Artisan;
